+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The following changes have been implemented but not released yet:
 
 ### New features
 
+- A new function `isValidVerifiablePresentation` has been added to the `verify` module, which verifies the validity of a VP using a verification service.
 - `getVerifiableCredentialApiConfiguration` now discovers the future-compatible
   specification-compliant endpoints, as well as the legacy endpoints.
 - A `query` function is added from the top-level export and the `@inrupt/solid-client-vc/query` submodule. It implements the Verifiable Presentation Request mechanism as described in https://w3c-ccg.github.io/vp-request-spec/. An important note is that we make the assumption that an endpoint supporting Verifiable Presentation Request is available at a /query path, which is outside of the VC API specification scope. This assumption is used to distinguish new endpoints vs legacy endpoints. Currently, only Query by Example VPRs are supported, which is similar to the legacy behavior in a lot of ways. The existing `getVerifiableCredentialAllFromShape` function now supports both legacy and VPR-compliant endpoints.